♻️ refactor :  특정 페이지에 종속된 컴포넌트는 해당 page.tsx와 같은 depth에 작성한다.
