+++ src/app/(Main)/talk/TalkSection.tsx
