+++ src/app/auth/AuthComponent.tsx
