--- src/app/auth/components/index.ts
@@ -1,1 +1,0 @@
-export { default as AuthComponent } from './AuthComponent';
