--- src/app/auth/withdraw/components/index.ts
@@ -1,1 +1,0 @@
-export { default as WithdrawComponent } from './WithdrawComponent';
