+++ src/app/auth/withdraw/page.tsx
@@ -1,3 +1,5 @@
+import WithdrawComponent from './WithdrawComponent';
+
 export default function Page() {
-  return null;
+  return <WithdrawComponent />;
 }
